feat: gradle 키 정리
