+++ app/build.gradle.kts
@@ -1,5 +1,4 @@
-import com.android.build.gradle.internal.cxx.configure.gradleLocalProperties
-import java.util.Properties
+/** See AndroidApplicationConventionPlugin.kt */
 
 plugins {
     alias(libs.plugins.acon.android.application)
@@ -9,32 +8,8 @@ plugins {
     alias(libs.plugins.acon.android.library.naver.map)
 }
 
-val localProperties = Properties().apply {
-    load(project.rootProject.file("local.properties").inputStream())
-}
-
 android {
     namespace = "com.acon.acon"
-
-    defaultConfig {
-        manifestPlaceholders += mapOf()
-        manifestPlaceholders["naverClientId"] = getPropertyKey("naver_client_id")
-        buildConfigField(
-            "String",
-            "NAVER_CLIENT_ID",
-            "String.valueOf(\"${localProperties["naver_client_id"]}\")"
-        )
-        buildConfigField("String", "BASE_URL", "String.valueOf(\"${localProperties["BASE_URL"]}\")")
-    }
-
-    signingConfigs {
-        create("release") {
-            storeFile = file(localProperties["storePath"].toString())
-            storePassword = localProperties["storePassword"].toString()
-            keyAlias = localProperties["keyAlias"].toString()
-            keyPassword = localProperties["keyPassword"].toString()
-        }
-    }
 
     buildTypes {
         release {
@@ -45,44 +20,24 @@ android {
                 "proguard-rules.pro"
             )
             signingConfig = signingConfigs.getByName("release")
-
-            manifestPlaceholders["naverClientId"] = getPropertyKey("naver_client_id")
-            buildConfigField(
-                "String",
-                "NAVER_CLIENT_ID",
-                "String.valueOf(\"${localProperties["naver_client_id"]}\")"
-            )
-            buildConfigField(
-                "String",
-                "BASE_URL",
-                "String.valueOf(\"${localProperties["BASE_URL"]}\")"
-            )
         }
     }
 }
 
-fun getPropertyKey(propertyKey: String): String {
-    val nullableProperty: String? =
-        gradleLocalProperties(rootDir, providers).getProperty(propertyKey)
-    return nullableProperty ?: "null"
-}
-
 dependencies {
 
-    implementation(project(":core:designsystem"))
-    implementation(project(":core:utils:feature"))
-    implementation(project(":core:map"))
-
-    implementation(project(":domain"))
-    implementation(project(":data"))
-
-    implementation(project(":feature:signin"))
-    implementation(project(":feature:spot"))
-    implementation(project(":feature:onboarding"))
-    implementation(project(":feature:areaverification"))
-    implementation(project(":feature:upload"))
-    implementation(project(":feature:settings"))
-    implementation(project(":feature:profile"))
+    implementation(projects.core.designsystem)
+    implementation(projects.core.utils.feature)
+    implementation(projects.core.map)
+    implementation(projects.domain)
+    implementation(projects.data)
+    implementation(projects.feature.signin)
+    implementation(projects.feature.spot)
+    implementation(projects.feature.onboarding)
+    implementation(projects.feature.areaverification)
+    implementation(projects.feature.upload)
+    implementation(projects.feature.settings)
+    implementation(projects.feature.profile)
 
     implementation(libs.play.services.location)
