+++ build-logic/convention/src/main/kotlin/utils/ProjectExtensions.kt
@@ -1,13 +1,11 @@
 package utils
 
 import com.android.build.api.dsl.CommonExtension
-import com.android.build.api.dsl.LibraryExtension
-import com.android.build.api.dsl.LibraryVariantDimension
+import com.android.build.gradle.internal.cxx.configure.gradleLocalProperties
 import org.gradle.api.JavaVersion
 import org.gradle.api.Project
 import org.gradle.api.artifacts.VersionCatalog
 import org.gradle.api.artifacts.VersionCatalogsExtension
-import org.gradle.api.artifacts.dsl.DependencyHandler
 import org.gradle.kotlin.dsl.getByType
 import org.gradle.kotlin.dsl.withType
 import org.jetbrains.kotlin.gradle.dsl.JvmTarget
@@ -15,6 +13,12 @@ import org.jetbrains.kotlin.gradle.tasks.KotlinJvmCompile
 
 val Project.catalog: VersionCatalog
     get() = extensions.getByType<VersionCatalogsExtension>().named("libs")
+
+internal fun Project.getPropertyKey(propertyKey: String): String {
+    val nullableProperty: String? =
+        gradleLocalProperties(rootDir, providers).getProperty(propertyKey)
+    return nullableProperty ?: "null"
+}
 
 internal fun Project.configureKotlinAndroid(
     commonExtension: CommonExtension<*, *, *, *, *, *>
